requirements_competition.txt
# Dépendances pour le générateur de compétition
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
reportlab>=4.0.0
tqdm>=4.66.0

# Dépendances déjà présentes (pour compatibilité)
huggingface_hub
datasets
pandas
